# bin/main/static/openapi.yaml
openapi: 3.0.3
info:
  title: Order Service API
  description: This is the REST API for Order Service
  version: v0.0.1
  license:
    name: Apache 2.0
externalDocs:
  description: You can refer to the Order Service Wiki Documentation
  url: https://order-service-dummy-url.com/docs
servers:
  - url: http://localhost:8081
    description: Local development server
paths:
  /api/order:
    post:
      summary: Place a new order
      description: Create a new order in the system
      operationId: placeOrder
      requestBody:
        description: Order information to be placed
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderRequest'
      responses:
        200:
          description: Order placed successfully
          content:
            text/plain:
              schema:
                type: string
                example: "Order placed successfully"
        400:
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        404:
          description: Products not found in inventory
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    OrderRequest:
      type: object
      required:
        - skuCode
        - price
        - quantity
      properties:
        skuCode:
          type: string
          example: "IPHONE13-128"
        price:
          type: number
          format: decimal
          example: 999.99
        quantity:
          type: integer
          format: int32
          example: 1
    OrderResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        orderNumber:
          type: string
          example: "f9b9a8b7-c6d5-4e3f-2g1h-0i9j8k7l6m5n"
        skuCode:
          type: string
          example: "IPHONE13-128"
        price:
          type: number
          format: decimal
          example: 999.99
        quantity:
          type: integer
          format: int32
          example: 1
        status:
          type: string
          enum: [PENDING, CONFIRMED, PROCESSING, SHIPPED, DELIVERED, CANCELLED, FAILED]
          example: "PENDING"
        createdAt:
          type: string
          format: date-time
          example: "2025-01-15T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-01-15T10:30:00Z"
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2023-05-20T14:30:15.123Z"
        status:
          type: integer
          example: 404
        error:
          type: string
          example: "Not Found"
        message:
          type: string
          example: "Product with Skucode: IPHONE13-128 is not in stock"
        path:
          type: string
          example: "/api/order" 